NIFI-1280: Addressed feedback from PR Review
