+++ nifi-nar-bundles/nifi-standard-bundle/nifi-standard-processors/src/test/java/org/apache/nifi/processors/standard/TestQueryFlowFile.java
@@ -17,32 +17,22 @@
 package org.apache.nifi.processors.standard;
 
 import java.io.IOException;
-import java.io.InputStream;
 import java.io.OutputStream;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.HashMap;
-import java.util.Iterator;
 import java.util.List;
-import java.util.Map;
 
 import org.apache.nifi.controller.AbstractControllerService;
-import org.apache.nifi.flowfile.FlowFile;
 import org.apache.nifi.logging.ComponentLog;
+import org.apache.nifi.processors.standard.util.record.MockRecordParser;
+import org.apache.nifi.processors.standard.util.record.MockRecordWriter;
 import org.apache.nifi.reporting.InitializationException;
-import org.apache.nifi.serialization.MalformedRecordException;
-import org.apache.nifi.serialization.RecordReader;
 import org.apache.nifi.serialization.RecordSetWriter;
 import org.apache.nifi.serialization.RecordSetWriterFactory;
-import org.apache.nifi.serialization.RowRecordReaderFactory;
-import org.apache.nifi.serialization.SimpleRecordSchema;
 import org.apache.nifi.serialization.WriteResult;
-import org.apache.nifi.serialization.record.MapRecord;
 import org.apache.nifi.serialization.record.Record;
-import org.apache.nifi.serialization.record.RecordField;
 import org.apache.nifi.serialization.record.RecordFieldType;
-import org.apache.nifi.serialization.record.RecordSchema;
 import org.apache.nifi.serialization.record.RecordSet;
 import org.apache.nifi.util.MockFlowFile;
 import org.apache.nifi.util.TestRunner;
@@ -272,121 +262,4 @@ public class TestQueryFlowFile {
 
     }
 
-    private static class MockRecordWriter extends AbstractControllerService implements RecordSetWriterFactory {
-        private final String header;
-
-        public MockRecordWriter(final String header) {
-            this.header = header;
-        }
-
-        @Override
-        public RecordSetWriter createWriter(final ComponentLog logger) {
-            return new RecordSetWriter() {
-                @Override
-                public WriteResult write(final RecordSet rs, final OutputStream out) throws IOException {
-                    out.write(header.getBytes());
-                    out.write("\n".getBytes());
-
-                    int recordCount = 0;
-                    final int numCols = rs.getSchema().getFieldCount();
-                    Record record = null;
-                    while ((record = rs.next()) != null) {
-                        recordCount++;
-                        int i = 0;
-                        for (final String fieldName : record.getSchema().getFieldNames()) {
-                            final String val = record.getAsString(fieldName);
-                            out.write("\"".getBytes());
-                            out.write(val.getBytes());
-                            out.write("\"".getBytes());
-
-                            if (i++ < numCols - 1) {
-                                out.write(",".getBytes());
-                            }
-                        }
-                        out.write("\n".getBytes());
-                    }
-
-                    return WriteResult.of(recordCount, Collections.emptyMap());
-                }
-
-                @Override
-                public String getMimeType() {
-                    return "text/plain";
-                }
-
-                @Override
-                public WriteResult write(Record record, OutputStream out) throws IOException {
-                    return null;
-                }
-            };
-        }
-    }
-
-    private static class MockRecordParser extends AbstractControllerService implements RowRecordReaderFactory {
-        private final List<Object[]> records = new ArrayList<>();
-        private final List<RecordField> fields = new ArrayList<>();
-        private final int failAfterN;
-
-        public MockRecordParser() {
-            this(-1);
-        }
-
-        public MockRecordParser(final int failAfterN) {
-            this.failAfterN = failAfterN;
-        }
-
-
-        public void addSchemaField(final String fieldName, final RecordFieldType type) {
-            fields.add(new RecordField(fieldName, type.getDataType()));
-        }
-
-        public void addRecord(Object... values) {
-            records.add(values);
-        }
-
-        @Override
-        public RecordReader createRecordReader(FlowFile flowFile, InputStream in, ComponentLog logger) throws IOException {
-            final Iterator<Object[]> itr = records.iterator();
-
-            return new RecordReader() {
-                private int recordCount = 0;
-
-                @Override
-                public void close() throws IOException {
-                }
-
-                @Override
-                public Record nextRecord() throws IOException, MalformedRecordException {
-                    if (failAfterN >= recordCount) {
-                        throw new MalformedRecordException("Intentional Unit Test Exception because " + recordCount + " records have been read");
-                    }
-                    recordCount++;
-
-                    if (!itr.hasNext()) {
-                        return null;
-                    }
-
-                    final Object[] values = itr.next();
-                    final Map<String, Object> valueMap = new HashMap<>();
-                    int i = 0;
-                    for (final RecordField field : fields) {
-                        final String fieldName = field.getFieldName();
-                        valueMap.put(fieldName, values[i++]);
-                    }
-
-                    return new MapRecord(new SimpleRecordSchema(fields), valueMap);
-                }
-
-                @Override
-                public RecordSchema getSchema() {
-                    return new SimpleRecordSchema(fields);
-                }
-            };
-        }
-
-        @Override
-        public RecordSchema getSchema(FlowFile flowFile) throws MalformedRecordException, IOException {
-            return null;
-        }
-    }
 }
