+++ nifi-nar-bundles/nifi-standard-services/nifi-record-serialization-services-bundle/nifi-record-serialization-services/src/main/java/org/apache/nifi/grok/GrokReader.java
@@ -91,7 +91,7 @@ public class GrokReader extends SchemaRegistryRecordReader implements RowRecordR
         }
 
         if (context.getProperty(PATTERN_FILE).isSet()) {
-            grok.addPatternFromFile(context.getProperty(PATTERN_FILE).getValue());
+            grok.addPatternFromFile(context.getProperty(PATTERN_FILE).evaluateAttributeExpressions().getValue());
         }
 
         grok.compile(context.getProperty(GROK_EXPRESSION).getValue());
